requirements.txt
mathutils==3.3.0
matplotlib==3.9.1
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
pyquaternion==0.9.9
scipy==1.14.0
tqdm==4.66.4